fix clang-tidy

+++ tasks/smetanin_d_gauss_vert_sch/tests/functional/main.cpp
@@ -25,9 +25,9 @@ bool SolveDenseSystem(std::vector<double> a, int n, std::vector<double> &x) {
   const double eps = 1e-10;
   for (int i = 0; i < n; ++i) {
     int pivot = i;
-    double best = std::fabs(a[static_cast<std::size_t>(i) * static_cast<std::size_t>(n + 1) + i]);
+    double best = std::fabs(a[(static_cast<std::size_t>(i) * static_cast<std::size_t>(n + 1)) + static_cast<std::size_t>(i)]);
     for (int rr = i + 1; rr < n; ++rr) {
-      double val = std::fabs(a[static_cast<std::size_t>(rr) * static_cast<std::size_t>(n + 1) + i]);
+      double val = std::fabs(a[(static_cast<std::size_t>(rr) * static_cast<std::size_t>(n + 1)) + static_cast<std::size_t>(i)]);
       if (val > best) {
         best = val;
         pivot = rr;
@@ -38,31 +38,31 @@ bool SolveDenseSystem(std::vector<double> a, int n, std::vector<double> &x) {
     }
     if (pivot != i) {
       for (int cj = i; cj <= n; ++cj) {
-        std::swap(a[static_cast<std::size_t>(i) * static_cast<std::size_t>(n + 1) + cj],
-                  a[static_cast<std::size_t>(pivot) * static_cast<std::size_t>(n + 1) + cj]);
+        std::swap(a[(static_cast<std::size_t>(i) * static_cast<std::size_t>(n + 1)) + static_cast<std::size_t>(cj)],
+            a[(static_cast<std::size_t>(pivot) * static_cast<std::size_t>(n + 1)) + static_cast<std::size_t>(cj)]);
       }
     }
     for (int rr = i + 1; rr < n; ++rr) {
-      double factor = a[static_cast<std::size_t>(rr) * static_cast<std::size_t>(n + 1) + i] /
-                      a[static_cast<std::size_t>(i) * static_cast<std::size_t>(n + 1) + i];
+      double factor = a[(static_cast<std::size_t>(rr) * static_cast<std::size_t>(n + 1)) + static_cast<std::size_t>(i)] /
+              a[(static_cast<std::size_t>(i) * static_cast<std::size_t>(n + 1)) + static_cast<std::size_t>(i)];
       if (std::fabs(factor) < eps) {
         continue;
       }
-      a[static_cast<std::size_t>(rr) * static_cast<std::size_t>(n + 1) + i] = 0.0;
+      a[(static_cast<std::size_t>(rr) * static_cast<std::size_t>(n + 1)) + static_cast<std::size_t>(i)] = 0.0;
       for (int cj = i + 1; cj <= n; ++cj) {
-        a[static_cast<std::size_t>(rr) * static_cast<std::size_t>(n + 1) + cj] -=
-            factor * a[static_cast<std::size_t>(i) * static_cast<std::size_t>(n + 1) + cj];
+        a[(static_cast<std::size_t>(rr) * static_cast<std::size_t>(n + 1)) + static_cast<std::size_t>(cj)] -=
+          factor * a[(static_cast<std::size_t>(i) * static_cast<std::size_t>(n + 1)) + static_cast<std::size_t>(cj)];
       }
     }
   }
 
   x.assign(static_cast<std::size_t>(n), 0.0);
   for (int i = n - 1; i >= 0; --i) {
-    double sum = a[static_cast<std::size_t>(i) * static_cast<std::size_t>(n + 1) + n];
+    double sum = a[(static_cast<std::size_t>(i) * static_cast<std::size_t>(n + 1)) + static_cast<std::size_t>(n)];
     for (int cj = i + 1; cj < n; ++cj) {
-      sum -= a[static_cast<std::size_t>(i) * static_cast<std::size_t>(n + 1) + cj] * x[static_cast<std::size_t>(cj)];
+      sum -= a[(static_cast<std::size_t>(i) * static_cast<std::size_t>(n + 1)) + static_cast<std::size_t>(cj)] * x[static_cast<std::size_t>(cj)];
     }
-    double diag = a[static_cast<std::size_t>(i) * static_cast<std::size_t>(n + 1) + i];
+    double diag = a[(static_cast<std::size_t>(i) * static_cast<std::size_t>(n + 1)) + static_cast<std::size_t>(i)];
     if (std::fabs(diag) < eps) {
       return false;
     }
